LPS-120702 Move DropZone component to same file as it's not being used elsewhere
