+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/DropZoneWithControls.js
@@ -19,8 +19,28 @@ import {
 	LayoutDataPropTypes,
 	getLayoutDataItemPropTypes,
 } from '../../../prop-types/index';
+import ManageAllowedFragmentButton from '../ManageAllowedFragmentButton';
 import Topper from '../Topper';
-import DropZone from './DropZone';
+
+const DropZone = React.forwardRef(({item}, ref) => {
+	return (
+		<div className="page-editor__drop-zone" ref={ref}>
+			<p>{Liferay.Language.get('drop-zone')}</p>
+
+			<p>
+				{Liferay.Language.get(
+					'fragments-and-widgets-for-pages-based-on-this-master-will-be-placed-here'
+				)}
+			</p>
+
+			<ManageAllowedFragmentButton item={item} />
+		</div>
+	);
+});
+
+DropZone.propTypes = {
+	item: getLayoutDataItemPropTypes().isRequired,
+};
 
 const DropZoneWithControls = React.forwardRef(({item, layoutData}, ref) => {
 	const [setRef, itemElement] = useSetRef(ref);
